revert: grpc response diff changes
